GenoDesigner-source/py/requirement.txt
Flask
Flask-APScheduler
Flask-Migrate
Flask-Login
Flask_Reuploaded
Flask-SQLAlchemy
flask-marshmallow
Bio
BCBio.GFF
openpyxl
xlsxwriter